chore: Test에 작성되었던 클래스 분리
