+++ app/src/main/java/woowacourse/omok/model/db/StoneContract.kt
@@ -1,0 +1,12 @@
+package woowacourse.omok.model.db
+
+import android.provider.BaseColumns
+
+object StoneContract {
+    object StoneEntry : BaseColumns {
+        const val TABLE_NAME = "stones"
+        const val COLUMN_NAME_X = "x"
+        const val COLUMN_NAME_Y = "y"
+        const val COLUMN_NAME_STONECOLOR = "stonecolor"
+    }
+}
